+++ fixtures/dom/src/__tests__/nested-act-test.js
@@ -8,10 +8,15 @@
  */
 
 let React;
-let TestUtils;
+let ReactDOM;
 let TestRenderer;
 
 global.__DEV__ = process.env.NODE_ENV !== 'production';
+
+jest.mock('react-dom', () =>
+  require.requireActual('react-dom/cjs/react-dom-testing.development.js')
+);
+// we'll replace the above with react/testing and react-dom/testing right before the next minor
 
 expect.extend(require('../toWarnDev'));
 
@@ -19,7 +24,7 @@ describe('unmocked scheduler', () => {
   beforeEach(() => {
     jest.resetModules();
     React = require('react');
-    TestUtils = require('react-dom/test-utils');
+    ReactDOM = require('react-dom');
     TestRenderer = require('react-test-renderer');
   });
 
@@ -33,7 +38,7 @@ describe('unmocked scheduler', () => {
     }
     // in legacy mode, this tests whether an act only flushes its own effects
     TestRenderer.act(() => {
-      TestUtils.act(() => {
+      ReactDOM.act(() => {
         TestRenderer.create(<Effecty />);
       });
       expect(log).toEqual([]);
@@ -42,7 +47,7 @@ describe('unmocked scheduler', () => {
 
     log = [];
     // for doublechecking, we flip it inside out, and assert on the outermost
-    TestUtils.act(() => {
+    ReactDOM.act(() => {
       TestRenderer.act(() => {
         TestRenderer.create(<Effecty />);
       });
@@ -59,7 +64,7 @@ describe('mocked scheduler', () => {
       require.requireActual('scheduler/unstable_mock')
     );
     React = require('react');
-    TestUtils = require('react-dom/test-utils');
+    ReactDOM = require('react-dom');
     TestRenderer = require('react-test-renderer');
   });
 
@@ -77,7 +82,7 @@ describe('mocked scheduler', () => {
     }
     // with a mocked scheduler, this tests whether it flushes all work only on the outermost act
     TestRenderer.act(() => {
-      TestUtils.act(() => {
+      ReactDOM.act(() => {
         TestRenderer.create(<Effecty />);
       });
       expect(log).toEqual([]);
@@ -86,7 +91,7 @@ describe('mocked scheduler', () => {
 
     log = [];
     // for doublechecking, we flip it inside out, and assert on the outermost
-    TestUtils.act(() => {
+    ReactDOM.act(() => {
       TestRenderer.act(() => {
         TestRenderer.create(<Effecty />);
       });
