+++ fixtures/dom/src/__tests__/wrong-act-test.js
@@ -12,12 +12,16 @@ let ReactDOM;
 let ReactART;
 let ARTSVGMode;
 let ARTCurrentMode;
-let TestUtils;
 let TestRenderer;
 let ARTTest;
 
 global.__DEV__ = process.env.NODE_ENV !== 'production';
 global.__EXPERIMENTAL__ = process.env.RELEASE_CHANNEL === 'experimental';
+
+jest.mock('react-dom', () =>
+  require.requireActual('react-dom/cjs/react-dom-testing.development.js')
+);
+// we'll replace the above with react/testing and react-dom/testing right before the next minor
 
 expect.extend(require('../toWarnDev'));
 
@@ -32,7 +36,6 @@ beforeEach(() => {
   ReactART = require('react-art');
   ARTSVGMode = require('art/modes/svg');
   ARTCurrentMode = require('art/modes/current');
-  TestUtils = require('react-dom/test-utils');
   TestRenderer = require('react-test-renderer');
 
   ARTCurrentMode.setCurrent(ARTSVGMode);
@@ -70,8 +73,8 @@ beforeEach(() => {
 });
 
 it("doesn't warn when you use the right act + renderer: dom", () => {
-  TestUtils.act(() => {
-    TestUtils.renderIntoDocument(<App />);
+  ReactDOM.act(() => {
+    ReactDOM.render(<App />, document.createElement('div'));
   });
 });
 
@@ -86,7 +89,7 @@ it('resets correctly across renderers', () => {
     React.useEffect(() => {}, []);
     return null;
   }
-  TestUtils.act(() => {
+  ReactDOM.act(() => {
     TestRenderer.act(() => {});
     expect(() => {
       TestRenderer.create(<Effecty />);
@@ -99,7 +102,7 @@ it('resets correctly across renderers', () => {
 it('warns when using the wrong act version - test + dom: render', () => {
   expect(() => {
     TestRenderer.act(() => {
-      TestUtils.renderIntoDocument(<App />);
+      ReactDOM.render(<App />, document.createElement('div'));
     });
   }).toWarnDev(["It looks like you're using the wrong act()"], {
     withoutStack: true,
@@ -113,7 +116,7 @@ it('warns when using the wrong act version - test + dom: updates', () => {
     setCtr = _setCtr;
     return ctr;
   }
-  TestUtils.renderIntoDocument(<Counter />);
+  ReactDOM.render(<Counter />, document.createElement('div'));
   expect(() => {
     TestRenderer.act(() => {
       setCtr(1);
@@ -123,7 +126,7 @@ it('warns when using the wrong act version - test + dom: updates', () => {
 
 it('warns when using the wrong act version - dom + test: .create()', () => {
   expect(() => {
-    TestUtils.act(() => {
+    ReactDOM.act(() => {
       TestRenderer.create(<App />);
     });
   }).toWarnDev(["It looks like you're using the wrong act()"], {
@@ -134,7 +137,7 @@ it('warns when using the wrong act version - dom + test: .create()', () => {
 it('warns when using the wrong act version - dom + test: .update()', () => {
   const root = TestRenderer.create(<App key="one" />);
   expect(() => {
-    TestUtils.act(() => {
+    ReactDOM.act(() => {
       root.update(<App key="two" />);
     });
   }).toWarnDev(["It looks like you're using the wrong act()"], {
@@ -151,15 +154,15 @@ it('warns when using the wrong act version - dom + test: updates', () => {
   }
   TestRenderer.create(<Counter />);
   expect(() => {
-    TestUtils.act(() => {
+    ReactDOM.act(() => {
       setCtr(1);
     });
   }).toWarnDev(["It looks like you're using the wrong act()"]);
 });
 
 it('does not warn when nesting react-act inside react-dom', () => {
-  TestUtils.act(() => {
-    TestUtils.renderIntoDocument(<ARTTest />);
+  ReactDOM.act(() => {
+    ReactDOM.render(<ARTTest />, document.createElement('div'));
   });
 });
 
@@ -171,7 +174,7 @@ it('does not warn when nesting react-act inside react-test-renderer', () => {
 
 it("doesn't warn if you use nested acts from different renderers", () => {
   TestRenderer.act(() => {
-    TestUtils.act(() => {
+    ReactDOM.act(() => {
       TestRenderer.create(<App />);
     });
   });
