+++ packages/react-dom/src/__tests__/ReactTestUtilsAct-test.js
@@ -722,6 +722,7 @@ function runActTests(label, render, unmount, rerender) {
 
     describe('suspense', () => {
       if (__DEV__ && __EXPERIMENTAL__) {
+        // todo - remove __DEV__ check once we start using testing builds
         it('triggers fallbacks if available', async () => {
           let resolved = false;
           let resolve;
