+++ packages/shared/forks/ReactFeatureFlags.test-renderer.www.js
@@ -44,6 +44,7 @@ export const disableUnstableRenderSubtreeIntoContainer = false;
 export const warnUnstableRenderSubtreeIntoContainer = false;
 export const disableUnstableCreatePortal = false;
 export const deferPassiveEffectCleanupDuringUnmount = false;
+export const isTestEnvironment = true; // this should probably *never* change
 
 // Only used in www builds.
 export function addUserTimingListener() {
