+++ packages/shared/forks/ReactFeatureFlags.testing.js
@@ -10,33 +10,27 @@
 import invariant from 'shared/invariant';
 
 import typeof * as FeatureFlagsType from 'shared/ReactFeatureFlags';
-import typeof * as FeatureFlagsShimType from './ReactFeatureFlags.native-fb';
+import typeof * as PersistentFeatureFlagsType from './ReactFeatureFlags.persistent';
 
-// Uncomment to re-export dynamic flags from the fbsource version.
-export const {
-  enableNativeTargetAsInstance,
-} = require('../shims/ReactFeatureFlags');
-
-// The rest of the flags are static for better dead code elimination.
+export const debugRenderPhaseSideEffectsForStrictMode = false;
 export const enableUserTimingAPI = __DEV__;
+export const warnAboutDeprecatedLifecycles = true;
+export const replayFailedUnitOfWorkWithInvokeGuardedCallback = false;
 export const enableProfilerTimer = __PROFILE__;
 export const enableSchedulerTracing = __PROFILE__;
 export const enableSuspenseServerRenderer = false;
 export const enableSelectiveHydration = false;
 export const enableChunksAPI = false;
+export const disableJavaScriptURLs = false;
+export const disableInputAttributeSyncing = false;
 export const exposeConcurrentModeAPIs = __EXPERIMENTAL__;
 export const warnAboutShorthandPropertyCollision = false;
 export const enableSchedulerDebugging = false;
-export const debugRenderPhaseSideEffectsForStrictMode = true;
-export const disableJavaScriptURLs = false;
-export const disableInputAttributeSyncing = false;
-export const replayFailedUnitOfWorkWithInvokeGuardedCallback = __DEV__;
-export const warnAboutDeprecatedLifecycles = true;
 export const enableDeprecatedFlareAPI = false;
 export const enableFundamentalAPI = false;
 export const enableScopeAPI = false;
 export const enableJSXTransformAPI = false;
-export const warnAboutUnmockedScheduler = true;
+export const warnAboutUnmockedScheduler = false;
 export const flushSuspenseFallbacksInTests = true;
 export const enableSuspenseCallback = false;
 export const warnAboutDefaultPropsOnFunctionComponents = false;
@@ -45,13 +39,14 @@ export const disableLegacyContext = false;
 export const disableSchedulerTimeoutBasedOnReactExpirationTime = false;
 export const enableTrainModelFix = true;
 export const enableTrustedTypesIntegration = false;
+export const enableNativeTargetAsInstance = false;
 export const disableCreateFactory = false;
 export const disableTextareaChildren = false;
 export const disableUnstableRenderSubtreeIntoContainer = false;
 export const warnUnstableRenderSubtreeIntoContainer = false;
 export const disableUnstableCreatePortal = false;
 export const deferPassiveEffectCleanupDuringUnmount = false;
-export const isTestEnvironment = false;
+export const isTestEnvironment = true;
 
 // Only used in www builds.
 export function addUserTimingListener() {
@@ -62,4 +57,4 @@ export function addUserTimingListener() {
 // eslint-disable-next-line no-unused-vars
 type Check<_X, Y: _X, X: Y = _X> = null;
 // eslint-disable-next-line no-unused-expressions
-(null: Check<FeatureFlagsShimType, FeatureFlagsType>);
+(null: Check<PersistentFeatureFlagsType, FeatureFlagsType>);
