+++ packages/shared/forks/ReactFeatureFlags.www.js
@@ -102,6 +102,8 @@ export const warnUnstableRenderSubtreeIntoContainer = false;
 
 export const disableUnstableCreatePortal = false;
 
+export const isTestEnvironment = false;
+
 // Flow magic to verify the exports of this file match the original version.
 // eslint-disable-next-line no-unused-vars
 type Check<_X, Y: _X, X: Y = _X> = null;
