+++ scripts/rollup/wrappers.js
@@ -1,25 +1,27 @@
 'use strict';
 
-const Bundles = require('./bundles');
+const {bundleTypes, moduleTypes} = require('./bundles');
 const reactVersion = require('../../package.json').version;
 
-const UMD_DEV = Bundles.bundleTypes.UMD_DEV;
-const UMD_PROD = Bundles.bundleTypes.UMD_PROD;
-const UMD_PROFILING = Bundles.bundleTypes.UMD_PROFILING;
-const NODE_DEV = Bundles.bundleTypes.NODE_DEV;
-const NODE_PROD = Bundles.bundleTypes.NODE_PROD;
-const NODE_PROFILING = Bundles.bundleTypes.NODE_PROFILING;
-const FB_WWW_DEV = Bundles.bundleTypes.FB_WWW_DEV;
-const FB_WWW_PROD = Bundles.bundleTypes.FB_WWW_PROD;
-const FB_WWW_PROFILING = Bundles.bundleTypes.FB_WWW_PROFILING;
-const RN_OSS_DEV = Bundles.bundleTypes.RN_OSS_DEV;
-const RN_OSS_PROD = Bundles.bundleTypes.RN_OSS_PROD;
-const RN_OSS_PROFILING = Bundles.bundleTypes.RN_OSS_PROFILING;
-const RN_FB_DEV = Bundles.bundleTypes.RN_FB_DEV;
-const RN_FB_PROD = Bundles.bundleTypes.RN_FB_PROD;
-const RN_FB_PROFILING = Bundles.bundleTypes.RN_FB_PROFILING;
+const {
+  UMD_DEV,
+  UMD_PROD,
+  UMD_PROFILING,
+  NODE_DEV,
+  NODE_PROD,
+  NODE_PROFILING,
+  FB_WWW_DEV,
+  FB_WWW_PROD,
+  FB_WWW_PROFILING,
+  RN_OSS_DEV,
+  RN_OSS_PROD,
+  RN_OSS_PROFILING,
+  RN_FB_DEV,
+  RN_FB_PROD,
+  RN_FB_PROFILING,
+} = bundleTypes;
 
-const RECONCILER = Bundles.moduleTypes.RECONCILER;
+const {RECONCILER} = moduleTypes;
 
 const license = ` * Copyright (c) Facebook, Inc. and its affiliates.
  *
