+++ scripts/shared/inlinedHostConfigs.js
@@ -11,6 +11,7 @@ module.exports = [
     shortName: 'dom',
     entryPoints: [
       'react-dom',
+      'react-dom/testing',
       'react-dom/unstable-fizz.node',
       'react-flight-dom-webpack/server.node',
       'react-flight-dom-webpack',
